fixing my github thing
