+++ src/main/java/me/tofpu/speedbridge/command/CommandHandler.java
@@ -3,12 +3,12 @@ package me.tofpu.speedbridge.command;
 import co.aikar.commands.BaseCommand;
 import co.aikar.commands.BukkitCommandManager;
 import me.tofpu.speedbridge.SpeedBridge;
-import me.tofpu.speedbridge.command.commands.AdminCommand;
-import me.tofpu.speedbridge.command.commands.MainCommand;
+import me.tofpu.speedbridge.command.sub.AdminCommand;
+import me.tofpu.speedbridge.command.sub.MainCommand;
 import me.tofpu.speedbridge.game.Game;
-import me.tofpu.speedbridge.game.controller.stage.SetupStage;
+import me.tofpu.speedbridge.game.controller.SetupStage;
 import me.tofpu.speedbridge.island.mode.Mode;
-import me.tofpu.speedbridge.island.mode.manager.ModeManager;
+import me.tofpu.speedbridge.island.mode.ModeManager;
 import me.tofpu.speedbridge.util.Util;
 
 public class CommandHandler {
