+++ src/main/java/me/tofpu/speedbridge/command/sub/MainCommand.java
@@ -1,12 +1,12 @@
-package me.tofpu.speedbridge.command.commands;
+package me.tofpu.speedbridge.command.sub;
 
 import co.aikar.commands.annotation.*;
-import me.tofpu.speedbridge.command.commands.base.BridgeBaseCommand;
+import me.tofpu.speedbridge.command.BridgeBaseCommand;
 import me.tofpu.speedbridge.data.file.path.Path;
-import me.tofpu.speedbridge.game.result.Result;
+import me.tofpu.speedbridge.game.Result;
 import me.tofpu.speedbridge.game.service.GameService;
 import me.tofpu.speedbridge.island.mode.Mode;
-import me.tofpu.speedbridge.island.mode.manager.ModeManager;
+import me.tofpu.speedbridge.island.mode.ModeManager;
 import me.tofpu.speedbridge.lobby.service.LobbyService;
 import me.tofpu.speedbridge.user.User;
 import me.tofpu.speedbridge.user.properties.UserProperties;
