+++ src/main/java/me/tofpu/speedbridge/data/adapter/LeaderboardTypeAdapter.java
@@ -4,7 +4,7 @@ import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonToken;
 import com.google.gson.stream.JsonWriter;
-import me.tofpu.speedbridge.lobby.leaderboard.data.BoardUser;
+import me.tofpu.speedbridge.lobby.leaderboard.BoardUser;
 
 import java.io.IOException;
 import java.util.ArrayList;
