+++ src/main/java/me/tofpu/speedbridge/data/adapter/UserAdapter.java
@@ -5,9 +5,9 @@ import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonToken;
 import com.google.gson.stream.JsonWriter;
 import me.tofpu.speedbridge.user.User;
-import me.tofpu.speedbridge.user.impl.UserImpl;
+import me.tofpu.speedbridge.user.UserImpl;
 import me.tofpu.speedbridge.user.properties.UserProperties;
-import me.tofpu.speedbridge.user.properties.timer.Timer;
+import me.tofpu.speedbridge.user.properties.Timer;
 
 import java.io.IOException;
 import java.util.UUID;
