+++ src/main/java/me/tofpu/speedbridge/data/file/path/PathType.java
@@ -1,4 +1,4 @@
-package me.tofpu.speedbridge.data.file.path.type;
+package me.tofpu.speedbridge.data.file.path;
 
 public enum PathType {
     SETTINGS, MESSAGES;
