+++ src/main/java/me/tofpu/speedbridge/data/file/path/Value.java
@@ -1,0 +1,5 @@
+package me.tofpu.speedbridge.data.file.path;
+
+public class Value<T> {
+    
+}
