--- src/main/java/me/tofpu/speedbridge/data/file/path/value/Value.java
@@ -1,5 +1,0 @@
-package me.tofpu.speedbridge.data.file.path.value;
-
-public class Value<T> {
-    
-}
