+++ src/main/java/me/tofpu/speedbridge/game/Game.java
@@ -5,17 +5,14 @@ import me.tofpu.speedbridge.command.CommandHandler;
 import me.tofpu.speedbridge.data.DataManager;
 import me.tofpu.speedbridge.game.controller.GameController;
 import me.tofpu.speedbridge.game.service.GameService;
-import me.tofpu.speedbridge.game.service.impl.GameServiceImpl;
+import me.tofpu.speedbridge.game.service.GameServiceImpl;
 import me.tofpu.speedbridge.island.service.IslandService;
-import me.tofpu.speedbridge.island.service.impl.IslandServiceImpl;
+import me.tofpu.speedbridge.island.service.IslandServiceImpl;
 import me.tofpu.speedbridge.lobby.service.LobbyService;
-import me.tofpu.speedbridge.lobby.service.impl.LobbyServiceImpl;
+import me.tofpu.speedbridge.lobby.service.LobbyServiceImpl;
 import me.tofpu.speedbridge.user.service.UserService;
-import me.tofpu.speedbridge.user.service.impl.UserServiceImpl;
+import me.tofpu.speedbridge.user.service.UserServiceImpl;
 import org.bstats.bukkit.Metrics;
-import org.bukkit.plugin.Plugin;
-
-import java.io.File;
 
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
