+++ src/main/java/me/tofpu/speedbridge/game/Result.java
@@ -1,4 +1,4 @@
-package me.tofpu.speedbridge.game.result;
+package me.tofpu.speedbridge.game;
 
 public enum Result {
     SUCCESS, DENY, FULL, NONE, INVALID_LOBBY, INVALID_ISLAND
