+++ src/main/java/me/tofpu/speedbridge/game/service/GameService.java
@@ -1,10 +1,10 @@
 package me.tofpu.speedbridge.game.service;
 
-import me.tofpu.speedbridge.game.result.Result;
+import me.tofpu.speedbridge.game.Result;
 import me.tofpu.speedbridge.island.Island;
 import me.tofpu.speedbridge.island.mode.Mode;
 import me.tofpu.speedbridge.user.User;
-import me.tofpu.speedbridge.user.properties.timer.Timer;
+import me.tofpu.speedbridge.user.properties.Timer;
 import org.bukkit.entity.Player;
 
 public interface GameService {
