+++ src/main/java/me/tofpu/speedbridge/game/service/GameServiceImpl.java
@@ -1,18 +1,17 @@
-package me.tofpu.speedbridge.game.service.impl;
+package me.tofpu.speedbridge.game.service;
 
 import com.google.common.collect.Lists;
 import me.tofpu.speedbridge.data.file.path.Path;
-import me.tofpu.speedbridge.game.result.Result;
-import me.tofpu.speedbridge.game.service.GameService;
+import me.tofpu.speedbridge.game.Result;
 import me.tofpu.speedbridge.island.Island;
 import me.tofpu.speedbridge.island.mode.Mode;
-import me.tofpu.speedbridge.island.mode.manager.ModeManager;
+import me.tofpu.speedbridge.island.mode.ModeManager;
 import me.tofpu.speedbridge.island.properties.twosection.TwoSection;
 import me.tofpu.speedbridge.island.service.IslandService;
 import me.tofpu.speedbridge.lobby.service.LobbyService;
 import me.tofpu.speedbridge.user.User;
 import me.tofpu.speedbridge.user.properties.UserProperties;
-import me.tofpu.speedbridge.user.properties.timer.Timer;
+import me.tofpu.speedbridge.user.properties.Timer;
 import me.tofpu.speedbridge.user.service.UserService;
 import me.tofpu.speedbridge.util.Cuboid;
 import me.tofpu.speedbridge.util.Util;
