+++ src/main/java/me/tofpu/speedbridge/island/mode/ModeManager.java
@@ -1,7 +1,6 @@
-package me.tofpu.speedbridge.island.mode.manager;
+package me.tofpu.speedbridge.island.mode;
 
 import com.github.requestpluginsforfree.config.ConfigAPI;
-import me.tofpu.speedbridge.island.mode.Mode;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.file.FileConfiguration;
