+++ src/main/java/me/tofpu/speedbridge/island/properties/twosection/IslandPoint.java
@@ -1,24 +1,20 @@
-package me.tofpu.speedbridge.island.properties.twosection.impl;
+package me.tofpu.speedbridge.island.properties.twosection;
 
-import me.tofpu.speedbridge.island.properties.twosection.TwoSection;
 import org.bukkit.Location;
 
-public class TwoSectionSimple implements TwoSection {
-    private String identifier = "";
+public class IslandPoint implements TwoSection {
+    private final String identifier;
 
     private Location sectionA;
     private Location sectionB;
 
-    public TwoSectionSimple() {
+    public IslandPoint() {
+        this.identifier = "point";
     }
 
     @Override
     public String getIdentifier() {
         return identifier;
-    }
-
-    public void setIdentifier(String identifier) {
-        this.identifier = identifier;
     }
 
     @Override
@@ -49,14 +45,5 @@ public class TwoSectionSimple implements TwoSection {
     @Override
     public boolean hasPointB() {
         return sectionB != null;
-    }
-
-    @Override
-    public String toString() {
-        return "TwoSectionSimple{" +
-                "identifier='" + identifier + '\'' +
-                ", sectionA=" + sectionA +
-                ", sectionB=" + sectionB +
-                '}';
     }
 }
