+++ src/main/java/me/tofpu/speedbridge/lobby/service/LobbyServiceImpl.java
@@ -1,10 +1,9 @@
-package me.tofpu.speedbridge.lobby.service.impl;
+package me.tofpu.speedbridge.lobby.service;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 import me.tofpu.speedbridge.lobby.leaderboard.Leaderboard;
-import me.tofpu.speedbridge.lobby.leaderboard.data.BoardUser;
-import me.tofpu.speedbridge.lobby.service.LobbyService;
+import me.tofpu.speedbridge.lobby.leaderboard.BoardUser;
 import org.bukkit.Location;
 
 import java.io.File;
